requirements.txt
Django==1.7
Django-Select2==4.2.2
django-bootstrap3==4.2.0
django-jsonview==0.3.0
django-sorting-bootstrap==1.0.2
django-su==0.4.2
yweather==0.1
psycopg2==2.5.4
djangorestframework==2.4.3
django-allauth==0.18.0
oauthlib==0.6.3
python-openid==2.2.5
requests==2.4.3
request-oauthlib==0.4.2
